fix(header): wire focus-mode hamburger to mobile drawer + drop dead useHeaderTier/rootRef
