+++ frontend/components/layout/useScrolled.ts
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+export function useScrolled(threshold = 0): boolean {
+  const [scrolled, setScrolled] = React.useState(false);
+  React.useEffect(() => {
+    const onScroll = () => setScrolled(window.scrollY > threshold);
+    onScroll();
+    window.addEventListener('scroll', onScroll, { passive: true });
+    return () => window.removeEventListener('scroll', onScroll);
+  }, [threshold]);
+  return scrolled;
+}
